About dataset_usedbikes
This dataset contains information about the used motorcycles. These motor cycles are mainly used in India and Asia. This data can be used for a lot of purposes such as price prediction to exemplify the use of linear regression in Machine Learning. 

The columns in the given dataset are as follows:
name selling price year seller type owner km driven ex showroom price